# backend/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        example: sth went wrong
        type: string
    type: object
...
  dto.RecommendedRule:
    properties:
      bank_code:
        type: string
      created_at:
        type: string
      product_id:
        type: string
      product_type:
        type: string
    type: object
  dto.RecommendedUpsertRequest:
    properties:
      bank_code:
        example: abank
        type: string
      product_id:
        example: prod-abank-card-001
        type: string
      product_type:
        example: credit_card
        type: string
    type: object
...
info:
  contact: {}
  description: API для аутентификации и пользователей.
  title: Multibank API
  version: "1.0"
paths:
...
  /admin/recommended-products:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Rule triplet to delete
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.RecommendedUpsertRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete recommended rule
      tags:
      - admin/products
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RecommendedRule'
            type: array
      security:
      - BearerAuth: []
      summary: List recommended product rules
      tags:
      - admin/products
    post:
      consumes:
      - application/json
      parameters:
      - description: Rule triplet
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.RecommendedUpsertRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upsert recommended rule
      tags:
      - admin/products
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"